requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.0
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
flake8==7.2.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycodestyle==2.13.0
pyflakes==3.3.2
sqlparse==0.5.3
tzdata==2025.2
